remove unnecessary fields in VO fold and edited classes
