+++ src/main/java/com/exadel/discountwebapp/role/vo/RoleBaseVO.java
@@ -1,6 +1,5 @@
 package com.exadel.discountwebapp.role.vo;
 
-import com.exadel.discountwebapp.user.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,8 +7,7 @@ import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 
 import java.io.Serializable;
-import java.time.LocalDateTime;
-import java.util.List;
+
 
 @Getter
 @Setter
@@ -17,8 +15,6 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public abstract class RoleBaseVO implements Serializable {
+
     protected String name;
-    protected List<User> users;
-    protected LocalDateTime created;
-    protected LocalDateTime modified;
 }
