+++ src/main/resources/application.yml
@@ -3,7 +3,7 @@ spring:
     url: jdbc:mysql://localhost:3306/discount_web_app
     driver-class-name: com.mysql.cj.jdbc.Driver
     username: root
-    password: root
+    password: pass123
   jpa:
     hibernate:
       ddl-auto: update
